requirements.txt
flask==3.1.0
supabase==2.13.0
python-dotenv==1.1.0
deepface==0.0.93
gunicorn==23.0.0
python-dateutil==2.9.0
opencv-python
numpy==1.26.4
tf-keras
Flask-Session==0.8.0
Flask-Limiter==3.9.0
